Trying to take logs under control
Added a custom Monitor to lower the logging level to INFO.
Unfortunately jetty logging is completely independent to this, and also
logs as DEBUG by default.
Finally solved it by taking advantage of logback, which was already
included in the project and controls the SLF4J logging facade used by jetty.

Decided to add code in the launcher itself. I think it's clean and
straightforward. Some things don't make sense as extensions.